Move frequency-space mds code into mds module
