+++ crypto/src/hash/mds/mod.rs
@@ -1,0 +1,10 @@
+// Copyright (c) Facebook, Inc. and its affiliates.
+//
+// This source code is licensed under the MIT license found in the
+// LICENSE file in the root directory of this source tree.
+
+//! An implementation of matrix-vector products for several MDS matrices.
+
+pub mod mds_f64_8x8;
+
+pub mod mds_f64_12x12;
