+++ crypto/src/hash/mod.rs
@@ -13,6 +13,8 @@ pub use blake::{Blake3_192, Blake3_256};
 mod sha;
 pub use sha::Sha3_256;
 
+mod mds;
+
 mod rescue;
 pub use rescue::{Rp62_248, Rp64_256, RpJive64_256};
